requirements.txt
streamlit
pandas
gspread
google-auth-oauthlib
folium
streamlit-folium
geopandas